requirements.txt
# Parallax Studio Requirements
# Target: macOS Apple Silicon (arm64)
# Python 3.11+

# UI Framework
PySide6==6.11.0

# Image Processing & Math
numpy==2.2.3
opencv-python-headless==4.11.0.86
Pillow==11.1.0

# Video & Animation Export
imageio[ffmpeg]==2.37.0
# Note: ffmpeg should also be installed via Homebrew for development
# brew install ffmpeg
